// x-pack/platform/plugins/shared/ingest_pipelines/public/types.d.ts
import type { ManagementSetup } from '@kbn/management-plugin/public';
import type { UsageCollectionSetup } from '@kbn/usage-collection-plugin/public';
import type { SharePluginStart, SharePluginSetup } from '@kbn/share-plugin/public';
import type { FileUploadPluginStart } from '@kbn/file-upload-plugin/public';
import type { ConsolePluginStart } from '@kbn/console-plugin/public';
import type { LicensingPluginStart } from '@kbn/licensing-plugin/public';
import type { ScopedHistory } from '@kbn/core/public';
export type { LicenseType, ILicense } from '@kbn/licensing-types';
export interface SetupDependencies {
    management: ManagementSetup;
    usageCollection: UsageCollectionSetup;
    share: SharePluginSetup;
}
export interface StartDependencies {
    share: SharePluginStart;
    fileUpload: FileUploadPluginStart;
    licensing?: LicensingPluginStart;
    console?: ConsolePluginStart;
}
export interface IngestPipelinesPluginStart {
    getIngestPipelineFlyoutComponent: (deps: {
        history: ScopedHistory<unknown>;
    }) => React.FC<IngestPipelineFlyoutProps>;
}
export interface IngestPipelineFlyoutProps {
    ingestPipelineName: string;
    onClose: () => void;
    reload: () => void;
}
export interface Config {
    enableManageProcessors: boolean;
}